Cast offsets to float

+++ masque/repetition.py
@@ -98,7 +98,7 @@ class GridRepetition:
 
         if val.size != 2:
             raise PatternError('Offset must be convertible to size-2 ndarray')
-        self._offset = val.flatten()
+        self._offset = val.flatten().astype(float)
 
     # dose property
     @property
